+++ components/water-mark/demo/cases/demo1.vue
@@ -21,7 +21,8 @@
 	</div>
 </template>
 
-<script>import {WaterMark} from 'al-mobile'
+<script>
+import {WaterMark} from 'al-mobile'
 
 export default {
   name: 'water-mark-demo',
@@ -33,16 +34,17 @@ export default {
     [WaterMark.name]: WaterMark,
   },
 }
-</script>
+
+</script>
 
 <style lang="stylus">
 .n22-example-child-water-mark-0
   .text-container
-    padding 32px
+    padding 16px
     background #FFF
   .text
-    margin-bottom 20px
+    margin-bottom 10px
     line-height 1.5
-    font-size 24px
+    font-size 12px
     text-indent 2em
 </style>